+++ package/server/src/ltn/ltns.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Ltn } from '../models/ltn.model';
+import dayjs from 'dayjs';
 
 @Injectable()
 export class LtnService {
@@ -11,7 +12,7 @@ export class LtnService {
 
   // 按照 boxId 分组
   async findAll({ start, end }): Promise<any> {
-    const ltns = await this.ltnModel.findAll();
+    const ltns = await this.ltnModel.findAll({ raw: true });
     const data = ltns.reduce((pre, cur) => {
       const boxId = cur.boxId;
       if (!pre[boxId]) {
@@ -37,6 +38,84 @@ export class LtnService {
       return pre;
     }, {});
     return { data };
+  }
+
+  async findAllCopy({ start, end }): Promise<any> {
+    const { data } = await this.findAll({ start, end });
+    const totalArr = Object.keys(data).reduce((pre, cur) => {
+      const count = data[cur].length;
+      pre.push(count);
+      return pre;
+    }, []);
+
+    const dataByDate = getDataByDate(data);
+
+    console.log('过滤后的数据data 按时间和 LTN 分类', totalArr, dataByDate);
+    return { totalArr, dataByDate };
+
+    function getDataByLtns(data) {
+      return Object.keys(data)
+        .map((it) => {
+          const ltns = data[it];
+          const idDesc = ltns.length
+            ? `
+    LTN${it} 【推荐做题时间 ${ltns[0].suggestTime} - ${ltns.length}题 实际做题时间： ？】
+    `
+            : '';
+          return (
+            idDesc +
+            ltns
+              .map(
+                (ltn) => `${ltn.title}
+    `,
+              )
+              .join('')
+          );
+        })
+        .join('');
+    }
+    function getDataByDate(data) {
+      const classifyByDate = {};
+      Object.keys(data).forEach((it) => {
+        const ltnData = data[it];
+        ltnData.forEach((ltn) => {
+          // 添加推荐做题时间
+          const suggestTime = dayjs(ltn.solveTime)
+            .add(ltn.boxId * 7, 'day')
+            .format('MM-DD');
+          const newOne = {
+            ...ltn,
+            suggestTime,
+          };
+          if (!classifyByDate[newOne.suggestTime]) {
+            classifyByDate[newOne.suggestTime] = [];
+          }
+          // 按照推荐做题时间分类
+          classifyByDate[newOne.suggestTime].push(newOne);
+        });
+      });
+      let res = '';
+      // 按照推荐做题时间排序
+      const sortDate = Object.keys(classifyByDate).sort((a, b) => {
+        return dayjs(a).isBefore(dayjs(b)) ? -1 : 1;
+      });
+
+      // 按照推荐做题时间分类
+      sortDate.forEach((it) => {
+        const ltns = classifyByDate[it];
+        const ltnsByBox = {};
+        ltns.forEach((ltn) => {
+          const boxId = ltn.boxId;
+          if (!ltnsByBox[boxId]) {
+            ltnsByBox[boxId] = [];
+          }
+          // 按照BOX分类
+          ltnsByBox[boxId].push(ltn);
+        });
+        res += getDataByLtns(ltnsByBox);
+      });
+      return res;
+    }
   }
 
   async updateBoxId(
